+++ service/history/timerQueueTaskExecutorBase.go
@@ -33,6 +33,7 @@ import (
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/backoff"
+	"go.temporal.io/server/common/definition"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/namespace"
@@ -300,7 +301,7 @@ func (t *timerQueueTaskExecutorBase) deleteWorkflowVisibility(
 		TransferTasks:    nil,
 		TimerTasks:       nil,
 		ReplicationTasks: nil,
-		VisibilityTasks: []persistence.Task{&persistence.DeleteExecutionVisibilityTask{
+		VisibilityTasks: []persistence.Task{&definition.DeleteExecutionVisibilityTask{
 			// TaskID is set by shard
 			VisibilityTimestamp: t.shard.GetTimeSource().Now(),
 			Version:             task.GetVersion(),
