+++ service/history/workflow/timer_sequence.go
@@ -37,7 +37,7 @@ import (
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/clock"
-	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/definition"
 	"go.temporal.io/server/common/primitives/timestamp"
 )
 
@@ -136,7 +136,7 @@ func (t *timerSequenceImpl) CreateNextUserTimer() (bool, error) {
 	if err := t.mutableState.UpdateUserTimer(timerInfo); err != nil {
 		return false, err
 	}
-	t.mutableState.AddTimerTasks(&persistence.UserTimerTask{
+	t.mutableState.AddTimerTasks(&definition.UserTimerTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: firstTimerTask.Timestamp,
 		EventID:             firstTimerTask.EventID,
@@ -182,7 +182,7 @@ func (t *timerSequenceImpl) CreateNextActivityTimer() (bool, error) {
 	if err != nil {
 		return false, err
 	}
-	t.mutableState.AddTimerTasks(&persistence.ActivityTimeoutTask{
+	t.mutableState.AddTimerTasks(&definition.ActivityTimeoutTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: firstTimerTask.Timestamp,
 		TimeoutType:         firstTimerTask.TimerType,
